chore: refactor command inits, add silencing of usage help text in case of error
