+++ cmd/cluster.go
@@ -221,5 +221,4 @@ func init() {
 	clusterCmd.AddCommand(clusterApplyCmd)
 	clusterCmd.AddCommand(clusterDestroyCmd)
 	clusterCmd.AddCommand(clusterTemplateCmd)
-	rootCmd.AddCommand(clusterCmd)
 }
