+++ cmd/completion.go
@@ -68,7 +68,3 @@ PowerShell:
 		}
 	},
 }
-
-func init() {
-	rootCmd.AddCommand(completionCmd)
-}
